requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.3
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
threadpoolctl==3.1.0
Werkzeug==2.3.4
zipp==3.15.0
